Fix package structure for java and scala sources
