+++ apps/src/main/java/org.seti/simulate/ArrayNoise.java
@@ -1,4 +1,5 @@
-package apps.simulate;
+package org.seti.simulate;
+
 import java.io.IOException;
 import java.nio.ByteBuffer;
