+++ apps/src/main/java/org.seti/simulate/FileNoise.java
@@ -1,4 +1,4 @@
-package apps.simulate;
+package org.seti.simulate;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
