+++ apps/src/main/java/org.seti/simulate/NoiseGenerator.java
@@ -1,4 +1,4 @@
-package apps.simulate;
+package org.seti.simulate;
 
 public class NoiseGenerator 
 {
