+++ spark/src/main/scala/org/seti/simulator/errors/Errors.scala
