+++ spark/src/main/scala/org/seti/simulator/signaldef/BrightPixel.scala
