+++ spark/src/main/scala/org/seti/simulator/signaldef/NarrowBand.scala
