+++ spark/src/main/scala/org/seti/simulator/signaldef/NarrowBandDRD.scala
