+++ spark/src/main/scala/org/seti/simulator/signaldef/PureNoise.scala
