+++ spark/src/main/scala/org/seti/simulator/signaldef/SignalDef.scala
