+++ spark/src/main/scala/org/seti/simulator/signaldef/SignalDefFactory.scala
