+++ spark/src/main/scala/org/seti/simulator/signaldef/SinePulsedNarrowBand.scala
