+++ spark/src/main/scala/org/seti/simulator/signaldef/SquarePulsedNarrowBand.scala
