+++ spark/src/main/scala/org/seti/simulator/signaldef/Squiggle.scala
