+++ spark/src/main/scala/org/seti/simulator/signaldef/SquiggleSinePulsedNarrowBand.scala
