+++ spark/src/main/scala/org/seti/simulator/signaldef/SquiggleSinePulsedNarrowBandDRD.scala
