+++ spark/src/main/scala/org/seti/simulator/signaldef/SquiggleSquarePulsedNarrowBand.scala
