+++ spark/src/main/scala/org/seti/simulator/signaldef/SquiggleSquarePulsedNarrowBandDRD.scala
