+++ spark/src/main/scala/org/seti/simulator/signaldef/TestSignal.scala
