+++ spark/src/main/scala/org/seti/simulator/utils/Utils.scala
